Added xslt-process-selected-style-frame to point to the last selected style frame.
Added xslt-process-speedbar-current-line-face, face for displaying the
current line in the speedbar.

(xslt-process-speedbar-change-style-frame): New function to select the
line of the selected style frame.

(xslt-process-speedbar-show-style-frames-stack): Invoke the
xslt-process-speedbar-change-style-frame function when an item in the
speedbar is clicked. Display the speedbar style frame item with
appropriate face. Pass the index of the frame to this function.